refactor: 분리된 테스트 클래스 Nested 클래스로 변경
